move Scopeguards, etc from renderer to util
