+++ tf2_bot_detector/ConsoleLog/ConsoleLines/LobbyStatusFailedLine.cpp
@@ -8,7 +8,7 @@
 #include "Log.h"
 #include "WorldState.h"
 
-#include <ScopeGuards.h>
+#include <Util/ScopeGuards.h>
 
 #include <regex>
 #include <sstream>
