+++ tf2_bot_detector/ConsoleLog/ConsoleLines/SuicideNotificationLine.cpp
@@ -13,7 +13,7 @@
 #include <mh/text/fmtstr.hpp>
 #include <mh/text/format.hpp>
 #include <mh/text/string_insertion.hpp>
-#include <ScopeGuards.h>
+#include <Util/ScopeGuards.h>
 
 #include <regex>
 #include <sstream>
