+++ tf2_bot_detector/UI/ImGui_TF2BotDetector.cpp
@@ -10,7 +10,7 @@
 #include "ReleaseChannel.h"
 
 #include <imgui_internal.h>
-#include <ScopeGuards.h>
+#include <Util/ScopeGuards.h>
 #include <mh/memory/cached_variable.hpp>
 #include <mh/text/string_insertion.hpp>
