+++ tf2_bot_detector/UI/ImGui_TF2BotDetector.h
@@ -2,8 +2,8 @@
 
 #include "ReleaseChannel.h"
 
-#include <ImGuiHelpers.h>
-#include <ScopeGuards.h>
+#include <Util/ImGuiHelpers.h>
+#include <Util/ScopeGuards.h>
 #include <imgui.h>
 #include <mh/raii/scope_exit.hpp>
 #include <mh/text/fmtstr.hpp>
