+++ tf2_bot_detector/UI/MainWindow.cpp
@@ -24,8 +24,8 @@
 #include "ConsoleLog/ConsoleLines/LobbyChangedLine.h"
 #include "ConsoleLog/ConsoleLines/EdictUsageLine.h"
 
-#include <ScopeGuards.h>
-#include <ImGuiHelpers.h>
+#include <Util/ScopeGuards.h>
+#include <Util/ImGuiHelpers.h>
 #include <imgui.h>
 #include <libzippp/libzippp.h>
 #include <misc/cpp/imgui_stdlib.h>
