+++ tf2_bot_detector/Util/ScopeGuards.cpp
