+++ tf2_bot_detector_renderer/sdl2opengl.cpp
@@ -57,7 +57,12 @@ TF2BotDetectorSDLRenderer::TF2BotDetectorSDLRenderer() : TF2BotDetectorRendererB
 	gl_context = SDL_GL_CreateContext(window);
 
 #ifdef IMGUI_USE_GLAD2
-	gladLoadGL([](const char* name) __declspec(noinline)
+	gladLoadGL([](const char* name) 
+#ifdef _WIN32
+	__declspec(noinline)
+#else
+	__attribute__((noinline))
+#endif
 	{
 		return reinterpret_cast<GLADapiproc>(SDL_GL_GetProcAddress(name));
 	});
